NIO - shutdown tests
